split moudle and fix regex bugs
