+++ .env.template
@@ -1,4 +1,4 @@
-# BASE_URL = "xxxxxxx"
-# AUTH_USER = "xxxxxxx"
-# AUTH_PASSWORD = "xxxxxxx"
-# BIJIBEN = 'xxxxxxx'
+# BASE_URL="https://dav.jianguoyun.com/dav/"
+# AUTH_USER = "20xxxxx957@qq.com"
+# AUTH_PASSWORD = "xxxxxxxxx"
+# BIJIBEN = 'web/test.md'
