+++ docker-compose-debug.yml
@@ -12,15 +12,11 @@ services:
       - 8889:8889
     volumes:
       - ./:/app
+      # - ./mysql/data:/var/lib/mysql
+      # - ./mysql/config:/etc/mysql/conf.d      
     depends_on:
       - mysql
-    # environment:
-    #   MYSQL_HOST: mysql
-    #   MYSQL_USER: root
-    #   MYSQL_PASSWORD: 123456
-    #   MYSQL_DB: diary_db
     restart: always
-    #command : sh -c "python /app/houtai.py"
     command: >
       sh -c "python -m debugpy --listen 0.0.0.0:8889 --wait-for-client  /app/houtai.py"
     networks:
