--- docker-compose-files/docker-compose1.yml
@@ -1,20 +1,0 @@
-services:
-  app:
-    build: ..
-    container_name: lzp_web
-    working_dir: /app
-    ports:
-      - 5001:5050
-    volumes:
-      # - ../:/app
-      - ../data:/app/data
-      - ../templates:/app/templates
-      - ../conf.py:/app/conf.py
-      - ../houtai.py:/app/houtai.py
-      
-      - ./data1/diary_1.csv:/app/diary.csv
-      - ./data1/ztb_1.csv:/app/ztb.csv
-      - ./data1/.env_1:/app/.env
-    restart: always
-    # command: sh -c "tail -f /dev/null"
-    command : sh -c "python /app/houtai.py"
