--- docker-compose-files/docker-compose2.yml
@@ -1,20 +1,0 @@
-services:
-  app2:
-    build: ..
-    container_name: lrx_web
-    working_dir: /app
-    ports:
-      - 5002:5050
-    volumes:
-      # - ../:/app
-      - ../data:/app/data
-      - ../templates:/app/templates
-      - ../conf.py:/app/conf.py
-      - ../houtai.py:/app/houtai.py
-      
-      - ./data2/diary_2.csv:/app/diary.csv
-      - ./data2/ztb_2.csv:/app/ztb.csv
-      - ./data2/.env_2:/app/.env
-    restart: always
-    # command: sh -c "tail -f /dev/null"
-    command : sh -c "python /app/houtai.py"
