--- docker-compose-files/init-files.sh
@@ -1,8 +1,0 @@
-mkdir -p data1
-mkdir -p data2
-touch data1/diary_1.csv
-touch data2/diary_2.csv
-touch data1/ztb_1.csv
-touch data2/ztb_2.csv
-touch data1/.env_1
-touch data2/.env_2
