--- docker-compose.yml
@@ -1,11 +1,0 @@
-services:
-  app:
-    build: .
-    working_dir: /app
-    ports:
-      - 5050:5050
-    volumes:
-      - ./:/app
-    restart: always
-    # command: sh -c "tail -f /dev/null"
-    command : sh -c "python /app/houtai.py"
